implemented get_miliseconds

+++ utils.h
@@ -6,7 +6,7 @@
 /*   By: bvelasco <bvelasco@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/19 14:04:57 by bvelasco          #+#    #+#             */
-/*   Updated: 2024/05/20 09:53:45 by bvelasco         ###   ########.fr       */
+/*   Updated: 2024/07/02 12:39:55 by bvelasco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@
  * 		to by the integer pointer.
  * Returns 0 if there are no errors; otherwise, returns the error code
  */
-int	ft_satoi(char *num, int *integer);
+short	ft_satoi(char *num, long *integer);
 #endif
